ci/openshift-ci: add updater for kernel modules in the initrd
When Kata is installed with `KATA_WITH_HOST_KERNEL=yes`, it is
configured to use the host kernel. The built initrd includes the kernel
modules of the buildroot (as hard-code in tests/versions.yaml), which
might not match the kernel running on the node at installation/test time, as
a result the VM crashes on boot. This added a handler on the installer
script which will update the initrd with the node's kernel modules in
case of mismatch.

It was also updated the kernel version in versions.yaml because the
CentOS 8 container now provides the EL 8.5 kernel (was EL 8.4).

Fixes #4163
Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>