k8s: add debug information to wait_pods_ready() in init.sh
It isn't uncommon the wait_pods_ready() in init.sh fail because some of the expected
system's pods didn't fully initialized. When that happens we don't have printed
further information about Kubernetes' pods. So this changed that method to print
debug information on fail and die with a user-friendly message.

Fixes #4247
Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>