k8s: retry the test environment initialization
The initialization of the test environment, which involves the start of
Kubernetes sometimes fail undetermistically. This added the ability to
retry the initialization on fail. Note that in case the clean up step
failed too, it will abort the operation because usually it is a sign
that something unexpected happened on the first initialization, so better
not to proceed to ensure the integrity of the test environment.

Fixes #4247
Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>