k8s: bring up kubelet with '--v 4' by default
Currently in integration/kubernetes/init.sh the kubelet is restart just
for the sake of passing `--v 4` (i.e. increase verbosity) and this has
been problematic because sometimes the daemon fail to join the cluster
again. When that error occur, the message is like:

 "Unable to connect to the server: net/http: TLS handshake timeout"

This changed the init script so that `--v 4` is passed to the kubelet when
`kubectl init` bring up the cluster; that eliminates the need for a
restart of kubelet and hopefully will turn the initialization more stable.

Fixes #4290
Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>