static-check: ignor http status of 429 when check URL
When check URL, "429" is sometimes returned.
"429 Too Many Requests" response status code indicates the user has
sent too many requests in a given amount of time. We can't predict
when the http server return this and it's not a fatal error. Also the
retry time lies in the "retry-after" of the header of response is too
long. So we can let it go and give a warning instead.

Fixes: #4404
Signed-off-by: Jianyong Wu <jianyong.wu@arm.com>