clh: Use kata-containers's static build script
Let's avoid code duplication and re-use here the very same script that's
used to build the kata-deploy binaries.

By doing this we:
* Reduce code duplication / complexity;
* Test the same build scripts used for kata-deploy on every single CI
  job;

One "drawback" is that we need to have the path to the cloud-hypervisor
binary, in the configuration file, adjusted to the new place where it's
installed. as the VMM installation path is changed, in the CI, from
`/usr/bin/cloud-hypervisor` to `/opt/kata/bin/cloud-hypervisor`.

Fixes: #4462

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>