ci: Fixed: convert nodename to lower case to match kubectl
`kubectl get nodes` returns all node names in lower case.
`hostname` returns the true host name, with capitalization honored.

This converts the hostname to lower case so that the master node
is successfully untainated for scheduling. Without this, the node
name is not recognized and the test fails with the error:

```
0/1 nodes are available: 1 node(s) had taint {node-role.kubernetes.io/master: }, that the pod didn't tolerate.
```

Fixes: #4502

Also added this to vfio.sh.

Signed-off-by: Agam Dua <agam_dua@apple.com>