ci: Added: tests for OCI prestart and poststart hooks.
Tests were created in a new `integration/ctr` directory,
which sets the stage for more work related to #4391.

The tests create a file on the guest vm which is shared
 with the container and adds the string `foo` to it.
A grep validates the string exists.
This was more helpful while debugging than a straight
presence check/stat of the file existing.

Fixes #4305.

Signed-off-by: Agam Dua <agam_dua@apple.com>