ci: add CC_SKOPEO_CRI_CONTAINERD CI job type
Currently the kata-agent integrates with skopeo to pull and verify
container images inside the guest. On the absence of skopeo, the agent
falls back to a (still) limited implementation. For example, without skopeo
the container policy file is not ignored, so images cannot be verified.

This adds the CC_SKOPEO_CRI_CONTAINERD job type which extends the
CC_CRI_CONTAINERD (added in commit 0678babef540884) configuration
so that skopeo is installed in the guest image. We will use that configuration
to test features which relies on skopeo.

Fixes #4572
Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>