ci: Add USE_DEVMAPPER flag
Currently selecting whether devmapper should (or should not) be used as
the containerd snapshotter is done based on the HYPERVISOR used with
Kata Containers, leading the only testing it when using Firecracker.

In order to prepare the ground for adding new jobs, let's ensure we
un-tie its configuration from the hypervisor being used.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>