scripts: Make die show stacktrace and other info under CI
Enhance the standard `die()` function to dump lots of information (in
YAML format) after displaying the error message when running under the
CI. The dump includes a full stacktrace and lots of environment details
to help with debugging.

This behaviour can also be forced by setting `KATA_TEST_VERBOSE=true`.

> **Note:**
>
> The conventional way of displaying this additional information is by
> defining a `trap` handler. However, that technique cannot be used here
> as this is effectively a library and the callers of this library can
> (and do) define their own `traps` (and only one `trap` can be defined
> per signal).

Fixes: #4732.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>