scripts: fix git security issue when install qemu for arm64
There is a new enhancement on security for git recently. It affects
git working on multi-user machines. eg. When working in git repo
whose user is non-root as a root, git may stop us to do something
affect other dirs out of this repo. We can add this repo explicitly
as a safe dir to avoid this issue.

see [1] to find more.

[1] https://github.blog/2022-04-12-git-security-vulnerability-announced/

Fixes: #4714
Signed-off-by: Jianyong Wu <jianyong.wu@arm.com>