container-manager: Fix docker checks
Now that the container manager script is using the central versions of
the info/warn/die functions, its behaviour needs to be changed slightly.
Previously, the `info_docker()` function was used in a strange way - it
always produced output, and always sent all output to stderr, even for
normal info messages. This is surprising and not Unix like so the script
now calls now calls `info_docker()` and checks its return
value. This means all output from `info_docker()` can go to standard
streams, as would be expected.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>