agent-shutdown: Fix command to wait for VM
In waiting for the VM to start, the test previously checked for ctr list
to report the container, but did not check its state. This sometimes
resulted in a race condition where the VM would be considered to be
running by the test even if it was not in the 'RUNNING' state, and the
the test would perform other actions on the VM, causing the test to
fail.

Update the command to specifically check for 'RUNNING' state in the
containerd task list to prevent failures.

Fixes #3878

Signed-off-by: Chelsea Mafrica <chelsea.e.mafrica@intel.com>