ci/lib.sh: fix sha256sum_from_files() not picking all files
The calc_qemu_files_sha256sum() passes a list of files and directories
to sha256sum_from_files() so that the sha256sum of QEMU build scripts is
calculated. However, sha256sum_from_files() is picking only the first parameter,
ie., in the end it calculates the hash of only one file/directory and,
therefore, the QEMU cache sha256sum is computed wrongly.

Fixes #4806
Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>