ci: change commit body line limit from 72 to 150
When this github action was ported over from kata-containers, I changed
the body length limit to 72 to match the community contribution
documentation. However, in the kata-containers repo it was changed back
to 150 after 72 was decided to be too short.

The community contribution documentation is being changed to 150, so
this should be changed to be consistent.

Fixes #4913

Signed-off-by: Derek Lee <derlee@redhat.com>