fc: fix uncompressed dir path
In the latest FC releases, the asset we download and decompress
has the following syntax:

release-vX.X.X-ARCH/firecracker-vX.X.X-ARCH

until v0.25 there was no ARCH in the folder name. Since we are updating
the supported FC version, this patch fixes the script that installs the
firecracker binary and the jailer.

Fixes: #4967
Signed-off-by: Anastassios Nanos <ananos@nubificus.co.uk>