tracing: Add checks for specific spans and ordering
Add check for spans in jaeger tracing output. Spans to be tested are
specified in test data in the tracing test script. Verify relative
ordering of spans to check that they occur in an expected order. For
example, test that some agent calls occur after specific runtime calls,
and that both occur under the root span.

Fixes: #4884

Signed-off-by: Chelsea Mafrica <chelsea.e.mafrica@intel.com>