metrics: Update ctr shim path for RunC
This PR removes the docker path that was used in kata 1.x for RunC
in the memory usage script. This PR updates the ctr shim path for
RunC in the memory usage script as now we are using ctr instead of
docker so we can measure correctly the pss memory.

Fixes #5011

Signed-off-by: Gabriela Cervantes <gabriela.cervantes.tellez@intel.com>