ci/install_runtime: build and install with kata-deploy
Changed the .ci/install_runtime.sh script so that it build the guest
image with kata-deploy's scripts then simply uncompress the tarball.

The binaries are now installed in /opt/confidential-containers/bin,
which is not in PATH, so they are linked in /usr/bin.

It doesn't need to change the configuration.toml as a new file for TDX
(configuration-qemu-tdx.toml) is created by the runtime package, so made
the replacement properly. Notice that if TDX and cloud hypervisor then
it still will `sed` the configuration.toml.

Fixes #4893
Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>