ci: Add `install_tvdf.sh`
Installing TDVF as part of the CI is needed in order to test QEMU + TDX
without depending on components internally provided by Intel.

Let's add a script to build and install tdvf into `/usr/share/tdvf`,
which takes advantage of the `cc-tdx-tdvf` target from the
`kata-deploy-binaries.sh`, which is the same one used for building the
binaries for the release, ensuring we're testing exactly what we're
shipping.

Fixes: #5061

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>