ci: Adjust TDVF location
TDVF, as it's been built as part of our CI, will be installed in:
`/usr/share/tdvf/`.

Let's adjust our scripts to point to that location.

Mind that in the near future we'll be able to drop those changes as
we'll be using a configuration file that's specific for TDX.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>