k8s: confidential: Increase the wait timeout to 60s
Let's double the wait timeout as I've noticed, in the internal TDX CI,
that pulling the image inside the guest may take longer than expected
due to our proxies.

In order to avoid random failures due to this, let's simply double the
amount of time and this will not have any side-effect on the other jobs.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>