ubuntu: Always install python3
We're facing the following issue as part of the CI:
```
Package python is not available, but is referred to by another package.
This may mean that the package is missing, has been obsoleted, or
is only available from another source
However the following packages replace it:
  python2-minimal python2 dh-python 2to3 python-is-python3
```

Considering that python3 is good enough for everything we need, let's
stick to it.

Fixes: #5091

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>