kubernetes/confidential: remove uneeded https handler
The pod configurations used on the confidential/kubernetes tests don't
set a DNS configuration. That is required by the configuration files
used on the cri-container tests only. As well as it doesn't need the
agent proxy configuration as Kubernetes handle that. So remove the uneeded
code that handle https proxy cases.

Fixes #4557
Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>