confidential/lib: allow to disable configure_cc_containerd()
The tests for confidential containers were implemented so that changes
for the containerd configuration file are required. For example, the
needed to pass the `cri_handler`. However, when those tests are ran from
the operator CI, the containerd is already proper configured. Moreover,
changing containerd's configuration can break it completely.

This commit changed the logic of configure_cc_containerd() so that if
`TESTS_CONFIGURE_CC_CONTAINERD=no` is exported then the function simply
does nothing.

Fixes #5077
Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>