tests: Pass --follow-symlinks when using `sed -i`
From https://linux.die.net/man/1/sed:
"""
-i[SUFFIX], --in-place[=SUFFIX]
edit files in place (makes backup if extension supplied). The default
operation mode is to break symbolic and hard links. This can be changed
with --follow-symlinks and --copy.
"""

This behaviour ends up breaking tests when we link the Kata Containers
configuration instead of copying it.

A big shout out to Chao Wu for pointing this out!

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>