ksm: Change how the shared_fs is set
Instead of using crudini, creating a backup file, and having to deal
with restoring it, let's simply `sed` our way to do the changes we need.

Hopefully this will be enough to get the monitor tests passing after
running the ksm ones, which is not happening with tests as they
currently are.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>