ci: replace $TRAVIS with $GITHUB_ACTION in setup
TRAVIS is not used for CI by kata-containers, and we're in the process of
removing TRAVIS variables from the environment. The GITHUB_ACTION variable is
defined by github actions by default ([1]) and it has the desired semantics
when checked here (only install a minimal set of extra packages).

I wanted to use $CI but that is also set on jenkins through
.ci/ci_job_flags.sh, and jenkins jobs do not work with the minimal setup.

[1]: https://docs.github.com/en/actions/learn-github-actions/variables#default-environment-variables)

Fixes: #3544
Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>