static-checks: Adjust sed expression allowing for checking more dockerfiles
So far most dockerfile templates have been on an ignore list, because the
dockerfile linter does not deal with the '@' symbol well. From reviewing the
dockerfiles, we use '@' in two places:
- to inject 'RUN' steps
- to inject image name/tags

The sed expression in the static-checks script handles the first one but not
the second one. Modify the sed expression to remove @ expressions at the
beginning of a line (handles case 1.) and drop the @ symbol if it appears
elsewhere in a line (handles case 2.).

With this change the ignore_files list can be dropped. Some of the dockerfiles
need slight tweaks to pass linting, but the required changes are reasonable and
will only be enforced next time a dockerfile is touched.

Fixes: #6472
Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>