kubernetes/confidential: Use resources from clone repo
setup_decryption_files_in_guest() current curls a file from github, while we
fetch the corresponding repo just a couple of lines later in
kubernetes_create_ssh_demo_pod(). Rewrite it to use checkout_doc_repo_dir() so
that the "depends-on" mechanism for changes to various repos works. It doesn't
work yet because checkout_doc_repo_dir() expects to make changes in the clone,
but this way there is a single point in the code where changes need to be made.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>