tests: Sstop running `make kubernetes` with containerd
Those tests are now covered as part of the GitHub actions we have, and
currently we're duplicating those (and also spending the resources
twice).

Let's stop running those for the cases we already cover with GitHub
actions, and update those little by little as we progress on the GHA
effort.

Fixes: #5643

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>