ci/openshift-ci: Add wrapper to ignore invariant failures
we are getting quite a lot of invariant tests failures, add this script
that will be used by "openshift/release" pipelines to post-process the
openshift-tests executions to ignore these for now (while reporting the
failures to xunit results).

Note the invariant tests are related on monitoring stability, which is
tricky when you replace containers with VMs. Anyway we still want to
investigate those but at this point we need to focus on the functional
tests more.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>